requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
supabase>=2.3.0
python-dotenv>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
python-telegram-bot>=21.0
rapidfuzz>=3.6.0
plotly>=5.18.0
numpy>=1.26.0
requests>=2.31.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
python-dateutil>=2.9.0
pytz>=2024.1
schedule>=1.2.0
aiohttp>=3.9.0
httpx>=0.27.0
